+++ projects/js-packages/image-guide/src/ui/Main.svelte
@@ -14,16 +14,16 @@
 	 * This onMount is triggered when the window loads
 	 * and the Image Guide UI is first
 	 */
-	onMount(() => {
-		stores.forEach(store => store.updateDimensions());
-	});
+	onMount( () => {
+		stores.forEach( store => store.updateDimensions() );
+	} );
 
-	function closeDetails(e) {
+	function closeDetails( e ) {
 		// Don't exit when hovering the Portal
 		if (
 			e.relatedTarget &&
 			// Don't exit when hovering the Popup
-			e.relatedTarget.classList.contains('keep-guide-open')
+			e.relatedTarget.classList.contains( 'keep-guide-open' )
 		) {
 			return;
 		}
@@ -31,34 +31,34 @@
 		show = false;
 	}
 
-	function getGuideSize(width = -1, height = -1): GuideSize {
-		if (width < 200 || height < 200) {
+	function getGuideSize( width = -1, height = -1 ): GuideSize {
+		if ( width < 200 || height < 200 ) {
 			return 'micro';
-		} else if (width < 400 || height < 400) {
+		} else if ( width < 400 || height < 400 ) {
 			return 'small';
 		}
 		return 'normal';
 	}
 
-	function toggleBackdrop(on = false) {
-		if (on) {
-			stores.forEach(store => store.node.classList.add('jetpack-boost-guide__backdrop'));
+	function toggleBackdrop( on = false ) {
+		if ( on ) {
+			stores.forEach( store => store.node.classList.add( 'jetpack-boost-guide__backdrop' ) );
 		} else {
-			stores.forEach(store => store.node.classList.remove('jetpack-boost-guide__backdrop'));
+			stores.forEach( store => store.node.classList.remove( 'jetpack-boost-guide__backdrop' ) );
 		}
 	}
 
 	// Use the first image available in the stores to determine the guide size
-	const sizeOnPage = stores[0].sizeOnPage;
-	$: size = getGuideSize($sizeOnPage.width, $sizeOnPage.height);
+	const sizeOnPage = stores[ 0 ].sizeOnPage;
+	$: size = getGuideSize( $sizeOnPage.width, $sizeOnPage.height );
 
-	$: toggleBackdrop(show !== false);
+	$: toggleBackdrop( show !== false );
 	let position = {
 		top: 0,
 		left: 0,
 	};
 
-	function hover(e: CustomEvent) {
+	function hover( e: CustomEvent ) {
 		const detail = e.detail;
 		const index = detail.index;
 		position = detail.position;
@@ -69,8 +69,9 @@
 	 * Only show image guide if at least one of the images
 	 * has a file size available.
 	 */
-	const hasItemsWithFileSize = derived(stores.map(s => s.fileSize), $fileSizes =>
-		$fileSizes.some(fileSize => fileSize.width !== -1 && fileSize.height !== -1)
+	const hasItemsWithFileSize = derived(
+		stores.map( s => s.fileSize ),
+		$fileSizes => $fileSizes.some( fileSize => fileSize.width !== -1 && fileSize.height !== -1 )
 	);
 </script>
 
@@ -94,20 +95,20 @@
 				Intentionally using only a single component here.
 				See <Popup> component source for details.
 			 -->
-			<Popup store={stores[show]} {size} {position} on:mouseleave={closeDetails} />
+			<Popup store={stores[ show ]} {size} {position} on:mouseleave={closeDetails} />
 		{/if}
 	</div>
 {/if}
 
 <style lang="scss">
-	:global(.jetpack-boost-guide) {
-		&:not(.relative) {
+	:global( .jetpack-boost-guide ) {
+		&:not( .relative ) {
 			position: absolute;
 			top: 0;
 			left: 0;
 		}
 	}
-	:global(.jetpack-boost-guide.relative) {
+	:global( .jetpack-boost-guide.relative ) {
 		position: relative;
 	}
 	.guide {
@@ -153,10 +154,10 @@
 		margin-bottom: 15px;
 	}
 
-	:global(.jetpack-boost-guide__backdrop) {
+	:global( .jetpack-boost-guide__backdrop ) {
 		transition:
 			opacity 0.2s ease-in-out,
 			filter 0.2s ease-in-out;
-		filter: brightness(0.3);
+		filter: brightness( 0.3 );
 	}
 </style>