+++ projects/github-actions/test-results-to-slack/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.2] - 2024-12-04
+### Added
+- Document required Slack scopes. [#39359]
+- Enable test coverage. [#39961]
+
 ## [0.3.1] - 2024-08-29
 ### Changed
 - Updated package dependencies. [#36097] [#36757] [#37706] [#38864]
@@ -60,6 +65,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove duplicated last run button for scheduled event notification
 - Remove duplicated last run button for workflow_run events
 
+[0.3.2]: https://github.com/Automattic/action-test-results-to-slack/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/Automattic/action-test-results-to-slack/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/Automattic/action-test-results-to-slack/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/Automattic/action-test-results-to-slack/compare/v0.2.0...v0.2.1